Struct mach_object::DyLibModule

pub struct DyLibModule {
    pub module_name: u32,
    pub iextdefsym: u32,
    pub nextdefsym: u32,
    pub irefsym: u32,
    pub nrefsym: u32,
    pub ilocalsym: u32,
    pub nlocalsym: u32,
    pub iextrel: u32,
    pub nextrel: u32,
    pub iinit_iterm: u32,
    pub ninit_nterm: u32,
    pub objc_module_info_addr: u32,
    pub objc_module_info_size: usize,
}

a module table entry

Fields
module_name: u32

the module name (index into string table)

iextdefsym: u32

index into externally defined symbols

nextdefsym: u32

number of externally defined symbols

irefsym: u32

index into reference symbol table

nrefsym: u32

number of reference symbol table entries

ilocalsym: u32

index into symbols for local symbols

nlocalsym: u32

number of local symbols

iextrel: u32

index into external relocation entries

nextrel: u32

number of external relocation entries

iinit_iterm: u32

low 16 bits are the index into the init section, high 16 bits are the index into the term section

ninit_nterm: u32

low 16 bits are the number of init section entries, high 16 bits are the number of term section entries

objc_module_info_addr: u32

for this module address of the start of the (OBJC,module_info) section

objc_module_info_size: usize

for this module size of the (OBJC,module_info) section