Struct mach_object::Section

pub struct Section {
    pub sectname: String,
    pub segname: String,
    pub addr: usize,
    pub size: usize,
    pub offset: u32,
    pub align: u32,
    pub reloff: u32,
    pub nreloc: u32,
    pub flags: SectionFlags,
    pub reserved1: u32,
    pub reserved2: u32,
    pub reserved3: u32,
}

A segment is made up of zero or more sections.

Non-MH_OBJECT files have all of their segments with the proper sections in each, and padded to the specified segment alignment when produced by the link editor. The first segment of a MH_EXECUTE and MH_FVMLIB format file contains the mach_header and load commands of the object file before its first section. The zero fill sections are always last in their segment (in all formats). This allows the zeroed segment padding to be mapped into memory where zero fill sections might be. The gigabyte zero fill sections, those with the section type S_GB_ZEROFILL, can only be in a segment with sections of this type. These segments are then placed after all other segments.

The MH_OBJECT format has all of its sections in one segment for compactness. There is no padding to a specified segment boundary and the mach_header and load commands are not part of the segment.

Sections with the same section name, sectname, going into the same segment, segname, are combined by the link editor. The resulting section is aligned to the maximum alignment of the combined sections and is the new section's alignment. The combined sections are aligned to their original alignment in the combined section. Any padded bytes to get the specified alignment are zeroed.

The format of the relocation entries referenced by the reloff and nreloc fields of the section structure for mach object files is described in the header file .

Fields
sectname: String

name of this section

segname: String

segment this section goes in

addr: usize

memory address of this section

size: usize

size in bytes of this section

offset: u32

file offset of this section

align: u32

section alignment (power of 2)

reloff: u32

file offset of relocation entries

nreloc: u32

number of relocation entries

flags: SectionFlags reserved1: u32

reserved (for offset or index)

reserved2: u32

reserved (for count or sizeof)

reserved3: u32

reserved